hevy_clone_backend/requirements.txt
fastapi[all] 
uvicorn[standard] 
sqlalchemy 
pydantic[email] 
pydantic-settings 
passlib[bcrypt] 
python-jose[cryptography] 
python-multipart 
pymysql